Fix dtype mismatch in dreamshaper_xl pipeline loading
The pretrained model has float16 weights that aren't fully converted by
torch_dtype=torch.float32 alone. Explicitly cast to float32 when moving
to CPU to prevent "expected m1 and m2 to have the same dtype" errors
during text encoder forward pass.

+++ dreamshaper_xl/pytorch/src/model_utils.py
@@ -26,8 +26,8 @@ def load_pipe(variant):
     pipe = DiffusionPipeline.from_pretrained(variant, torch_dtype=torch.float32)
     modules = [pipe.text_encoder, pipe.unet, pipe.text_encoder_2, pipe.vae]
 
-    # Move the pipeline to CPU
-    pipe.to("cpu")
+    # Move the pipeline to CPU and ensure float32 dtype
+    pipe.to("cpu", dtype=torch.float32)
 
     for module in modules:
         module.eval()
